Former-commit-id: 5bd46d21f74c2ea5a4dd899f70b5242d3d45c116

+++ framework/JPetEvent/JPetEvent.h
@@ -91,7 +91,7 @@ public:
 
   JPetHit fFirstHit;
   JPetHit fSecondHit;
-  float fNhits;
+  unsigned short int fNhits;
 
 };
